Rename InputControl to InputComponent for consistency
